DSODA-111 split stream-metadata and woms into two services
